app/requirements.txt
RPi.GPIO==0.7.1
python-barcode==0.13.1
psycopg2==2.9.1
python-dotenv==0.19.0
gpiozero==1.6.2
pillow==8.3.1
pydub==0.25.1
